+++ src/main/java/com/petshop/petopia/dto/response/banner/PetWithDiscount.java
@@ -1,6 +1,5 @@
 package com.petshop.petopia.dto.response.banner;
 
-import com.petshop.petopia.model.pet.Pet;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -16,6 +15,6 @@ public class PetWithDiscount {
     private Integer id;
     private String avatarUrl;
     private String name;
-    private Integer originalPrice;
-    private Integer discountedPrice;
+    private BigDecimal originalPrice;
+    private BigDecimal discountedPrice;
 }
