+++ src/main/java/com/petshop/petopia/dto/response/banner/ProductWithDiscount.java
@@ -1,9 +1,9 @@
 package com.petshop.petopia.dto.response.banner;
 
-import com.fasterxml.jackson.core.JsonToken;
-import com.petshop.petopia.model.product.Product;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -11,6 +11,6 @@ public class ProductWithDiscount {
     private Integer id;
     private String imageUrl;
     private String name;
-    private Integer originalPrice;
-    private Integer discountedPrice;
+    private BigDecimal originalPrice;
+    private BigDecimal discountedPrice;
 }
