+++ src/main/java/com/petshop/petopia/dto/response/cart/CartItemResponse.java
@@ -3,6 +3,8 @@ package com.petshop.petopia.dto.response.cart;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class CartItemResponse {
@@ -12,6 +14,6 @@ public class CartItemResponse {
     private Integer petId;
     private String petName;
     private Integer quantity;
-    private Integer price;
-    private Integer itemTotalPrice;
+    private BigDecimal price;
+    private BigDecimal itemTotalPrice;
 }
