+++ src/main/java/com/petshop/petopia/dto/response/order/OrderResponse.java
@@ -1,16 +1,18 @@
 package com.petshop.petopia.dto.response.order;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
 @Data
 public class OrderResponse {
     private Integer id;
-    private Long userId; // Sử dụng Long để phù hợp với kiểu dữ liệu của User
+    private Long userId;
     private List<OrderItemResponse> items;
     private Date orderDate;
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
     private String shippingAddress;
     private String phoneNumber;
     private PaymentResponse paymentResponse;
